% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/find_best_k.R
\name{find_best_k}
\alias{find_best_k}
\title{Finds the best value of k for the nearest neighbor classifier, judged by holdout accuracy}
\usage{
find_best_k(data, trueClasses, p = 2, k_seq = c(1, 2), holdout = 0.3,
  seed = NA)
}
\arguments{
\item{data}{A data frame or a matrix where rows are observations and
columns are features}

\item{trueClasses}{A vector with labels for each row in \code{data}}

\item{p}{Distance metric the k-NN Classifier should use, the value can be
either 1 (Manhattan Distance), 2(Euclidean Distance) or Inf (Chebyshev Distance).}

\item{k_seq}{a vector with the values for k to be evaluated}

\item{holdout}{the relative size of the holdout validation set - must be between 0 and 1}

\item{seed}{- optional; used for random shuffling, set for reproducible behavior}
}
\value{
a list with following objects: dataframe containing holdout and training accuracy,
the actual best value for k judged by holdout accuracy and a plot of K vs. Training/Holdout Accuracy
}
\description{
Finds the best value of k for the nearest neighbor classifier, judged by holdout accuracy
}
\examples{
# create artificial dataset
inputs  <- matrix(rnorm(200), ncol=2)
classes <- c(rep(0, 50), rep(1, 50))
k_seq <- c(1,3,5)
results <- find_best_k(inputs, classes,2, k_seq)
results$plot
}

